requirements.txt
numpy>=1.21.0,<2.0.0
fastapi==0.105.0
uvicorn==0.24.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
pillow==10.1.0
python-multipart==0.0.6